controller: Add controller to create issues
This controller established to request mappings.

If a GET request is received the controller will render a page with a
form that the user can use to create new GitHub issues.

On a POST request the content of the form is processed and currently
the content of the form is shown.